* changed signature of maximal_independent_set method (#2975)
* * changed signature of maximal_independent_set method
to accept a table of vertices and return a filtered table.

* changed BinaryHeap's tiebreaker function to return
a Long, which matches data type of indices returned by
Table.index()

* added selectGlobal method to Table which takes String*

* *edited signature of maximal_independent_set
to include a boolean and return a Table.

changes after rebase

fixed example

rebased

removed import statement

* addresses comments

+++ src/main/scala/is/hail/utils/BinaryHeap.scala
@@ -3,7 +3,7 @@ package is.hail.utils
 import scala.collection.mutable
 import scala.reflect.ClassTag
 
-class BinaryHeap[@specialized T: ClassTag](minimumCapacity: Int = 32, maybeTieBreaker: (T, T) => Int = null) {
+class BinaryHeap[@specialized T: ClassTag](minimumCapacity: Int = 32, maybeTieBreaker: (T, T) => Long = null) {
   private var ts: Array[T] = new Array[T](minimumCapacity)
   private var ranks: Array[Long] = new Array[Long](minimumCapacity)
   private val m: mutable.Map[T, Int] = new mutable.HashMap()
@@ -15,6 +15,7 @@ class BinaryHeap[@specialized T: ClassTag](minimumCapacity: Int = 32, maybeTieBr
   def size: Int = next
 
   def isEmpty: Boolean = next == 0
+
   def nonEmpty: Boolean = next != 0
 
   override def toString(): String =
